Major changes to convert remaining elements to use CustomTkinter
